+++ a.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-from lang_trans.japanese import hepburn
-
-ORIG = u'じゃしゃしんいっぱい'
-TRANS = 'jashashinippai'
-
-print hepburn.untrans(TRANS)
-print ORIG