+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.3.0
 
-* Add Japanese
+* Add Japanese hepburn
+* Modify the API for simpler use
 
 ## 0.2.2